Login is done if another Language is added all of the GUI Items need to fully added in Labels. Otherwise the Login will crash when other language is selected.
Main not ready for final design but thats quite far away yet.

Main needs to be selecting the entire Database from Server or Excel Sheet.
Main needs to make user select between these two optiones
 -> Database or
 -> Excel sheet

For internation or regional events global database should be created and selected from the GUI automatically

All events (local, national and regional) need to be displayed and then the action items for every sec. should be listed in a window.

Reminder functionality needs to be added in the Main GUI.
Either by sending the message to the smartphone or by email or on the screen. Maybe with an ICS file that saves event on smartphone calendar.

.....